Changes after feedback

+++ app/src/main/java/fr/exaltit/fizz_buzz/domain/model/FizzBuzzData.kt
@@ -1,9 +1,9 @@
 package fr.exaltit.fizz_buzz.domain.model
 
 data class FizzBuzzData(
-	var firstMultiple: Int,
-	var secondMultiple: Int,
-	var limit: Int,
+	var firstMultiple: Long,
+	var secondMultiple: Long,
+	var limit: Long,
 	var firstWord: String,
 	var secondWord: String
 )
